+++ libs/bpm-core-client/src/lib/form/index.ts
@@ -1,2 +1,3 @@
 export * from './form-api';
+export * from './form-data-source-api';
 export * from './form-rendering';